requirements.txt
asgiref==3.5.0
Django==4.0.2
django-cors-headers==3.11.0
django-filter==21.1
djangorestframework==3.13.1
importlib-metadata==4.11.1
Markdown==3.3.6
pytz==2021.3
sqlparse==0.4.2
zipp==3.7.0

pip~=22.0.4
wheel~=0.36.2
setuptools~=57.0.0
requests~=2.27.1
djangorestframework-jwt~=1.11.0